+++ src/HDInsight/HDInsight/HDInsightCmdletBase.cs
@@ -15,7 +15,7 @@
 using Hyak.Common;
 using Microsoft.Azure.Commands.HDInsight.Models;
 using Microsoft.Azure.Commands.ResourceManager.Common;
-using Microsoft.Azure.Management.HDInsight;
+using Microsoft.Azure.Management.HDInsight.Models;
 using System;
 using System.Linq;